fix: background interaction modifier through struct
