+++ Example/BottomSheetExample/BottomSheetExampleApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct BottomSheetExampleApp: App {
     var body: some Scene {
         WindowGroup {
-            NavigationViewV2()
+            ExampleOverview()
         }
     }
 }
