+++ Sources/BottomSheet/BottomSheet.swift
@@ -1,158 +1,157 @@
 //
-//  BottomSheet.swift
+//  SwiftUIView.swift
+//  
 //
-//
-//  Created by Wouter van de Kamp on 26/11/2022.
+//  Created by Wouter van de Kamp on 06/01/2024.
 //
 
 import SwiftUI
 
-struct SheetPlus<HContent: View, MContent: View, Background: View>: ViewModifier {
+struct SheetPlus<HContent: View, MContent: View>: ViewModifier {
     @Binding private var isPresented: Bool
-    
-    @State private var translation: CGFloat = 0
-    @State private var sheetConfig: SheetPlusConfig?
-    @State private var showDragIndicator: VisibilityPlus?
-    @State private var allowBackgroundInteraction: PresentationBackgroundInteractionPlus?
-    @State private var isInteractiveDismissDisabled = true
 
+    @State private var translation: CGFloat = 0
     @State private var newValue = 0.0
     @State private var startTime: DragGesture.Value?
-    
-    @State private var detents: Set<PresentationDetent> = []
+
+    @State private var sheetConfig: SheetPlusConfig?
     @State private var limits: (min: CGFloat, max: CGFloat) = (min: 0, max: 0)
 
-    @State private var initialSelectedDetent: PresentationDetent? = nil
-    @State private var isDismissed = false
+    @State private var detents: Set<PresentationDetent> = []
 
+    @State private var showDragIndicator: VisibilityPlus?
+    @State private var cornerRadius: CGFloat?
+    @State private var background: EquatableBackground?
+    @State private var isInteractiveDismissDisabled = true
+
+    @State private var backgroundInteractionDetentLimit: CGFloat?
+    @State private var isBackgroundInteractionEnabled = true
+
+    @State private var viewWillPresent = false
+
+    let animationCurve: SheetAnimation
+    let onDrag: (CGFloat) -> Void
     let mainContent: MContent
     let headerContent: HContent
-    let animationCurve: SheetAnimation
     let onDismiss: () -> Void
-    let onDrag: (CGFloat) -> Void
-    let background: Background
-    
+
     init(
         isPresented: Binding<Bool>,
         animationCurve: SheetAnimation,
-        background: Background,
         onDismiss: @escaping () -> Void,
-        onDrag: @escaping (CGFloat) -> Void,
+        onDrag: @escaping (CGFloat) -> Void = { _ in },
         @ViewBuilder hcontent: () -> HContent,
         @ViewBuilder mcontent: () -> MContent
     ) {
         self._isPresented = isPresented
-        
         self.animationCurve = animationCurve
-        self.background = background
+
         self.onDismiss = onDismiss
         self.onDrag = onDrag
-        
+
         self.headerContent = hcontent()
         self.mainContent = mcontent()
     }
-    
+
+    func isInteractionEnabled() -> Bool {
+        return true
+    }
+
     func body(content: Content) -> some View {
-        ZStack() {
+        ZStack {
             content
-//                .allowsHitTesting(allowBackgroundInteraction == .disabled ? false : true)
-                .onChange(of: isPresented) { newValue in
-                    guard let initialSelectedDetent = initialSelectedDetent else { return }
-
-                    if newValue == true {
-                        isDismissed = false
-                        sheetConfig?.selectedDetent = initialSelectedDetent
-                    }
-
-                    if newValue == false {
-                        translation = 0
-                        sheetConfig?.selectedDetent = PresentationDetent.height(.zero)
+                .allowsHitTesting(!isPresented || isBackgroundInteractionEnabled)
+                .zIndex(1)
+                .onChange(of: isPresented) { value in
+                    if (value == true) {
+                        viewWillPresent = value
+                    } else {
+                        translation = .zero
                     }
                 }
 
-            if !isDismissed {
-                GeometryReader { geometry in
-                    VStack(spacing: 0) {
-                        // If / else statement here breaks the animation from the bottom level
-                        // Might want to see if we can refactor the top level animation a bit
-                        DragIndicator(
-                            translation: $translation,
-                            detents: detents
-                        )
-                            .frame(height: showDragIndicator == .visible ? 22 : 0)
-                            .opacity(showDragIndicator == .visible ? 1 : 0)
+            if viewWillPresent {
+                // VStack to stick the sheet to the bottom
+                VStack(spacing: 0) {
+                    Spacer()
 
-                        headerContent
+                    // VStack to keep the drag indicator and header at the top of the card
+                    VStack(spacing: 0) {
+                        // Geometry reader keeps the content outside the header vertically centered
+//                        GeometryReader { geometry in
+                            // Holds the header customization
+                            VStack(spacing: 0) {
+                                DragIndicator(
+                                    translation: $translation,
+                                    detents: detents
+                                )
+                                .frame(height: showDragIndicator == .hidden ? 0 : 22)
+                                .opacity(showDragIndicator == .hidden ? 0 : 1)
+
+                                headerContent
+                            }
+//                            .frame(width: geometry.size.width)
                             .contentShape(Rectangle())
                             .gesture(
                                 DragGesture(coordinateSpace: .global)
                                     .onChanged { value in
                                         translation -= value.location.y - value.startLocation.y - newValue
                                         newValue = value.location.y - value.startLocation.y
-                                        
+
                                         if startTime == nil {
                                             startTime = value
                                         }
                                     }
                                     .onEnded { value in
-                                        // Reset the distance on release so we start with a
-                                        // clean translation next time
-                                        newValue = 0
-                                        
-                                        // Calculate velocity based on pt/s so it matches the UIPanGesture
-                                        let distance: CGFloat = value.translation.height
-                                        let time: CGFloat = value.time.timeIntervalSince(startTime!.time)
-                                        
-                                        let yVelocity: CGFloat = -1 * ((distance / time) / 1000)
-                                        startTime = nil
-
-                                        if let result = snapBottomSheet(
-                                            translation,
-                                            detents,
-                                            yVelocity,
-                                            isInteractiveDismissDisabled
-                                        ) {
-                                            translation = result.size
-                                            sheetConfig?.selectedDetent = result
-                                        }
+                                        onDragEnded(with: value)
                                     }
                             )
+//                        }
 
-                        UIScrollViewWrapper(
-                            translation: $translation,
-                            preferenceKey: $sheetConfig,
-                            isInteractiveDismissDisabled: $isInteractiveDismissDisabled,
-                            limits: limits,
-                            detents: detents
-                        ) {
-                            mainContent
-                                .frame(width: geometry.size.width)
+//                        Spacer()
+
+                        GeometryReader { geometry in
+                            UIScrollViewWrapper(
+                                translation: $translation,
+                                preferenceKey: $sheetConfig,
+                                isInteractiveDismissDisabled: $isInteractiveDismissDisabled,
+                                limits: limits,
+                                detents: detents
+                            ) {
+                                mainContent
+                                    .frame(width: geometry.size.width)
+                            }
                         }
+
+//                        Spacer()
                     }
-                    .background(background)
-                    .frame(height:
-                            (limits.max - geometry.safeAreaInsets.top) > 0
-                                ? limits.max - geometry.safeAreaInsets.top
-                                : limits.max
+                    .background(
+                        ZStack() {
+                            background?.view
+                                .cornerRadius(cornerRadius ?? 0)
+                        }
                     )
-                    .onChange(of: translation) { newValue in
-                        // Small little hack to make the iOS scroll behaviour work smoothly
-                        if limits.max == 0 { return }
+                    .frame(
+                        width: UIScreen.main.bounds.width,
+                        height: translation
+                    )
+                    .clipped()
+                    .onChange(of: translation) { value in
+                        let minLimit = isInteractiveDismissDisabled && isPresented ? limits.min : .zero
+                        translation = min(limits.max, max(value, minLimit))
 
-                        let minValue = isInteractiveDismissDisabled && isPresented ? limits.min : 0
-                        translation = min(limits.max, max(newValue, minValue))
-
-//                        currentGlobalTranslation = translation
+                        if let interactionDetent = backgroundInteractionDetentLimit {
+                            isBackgroundInteractionEnabled = translation < interactionDetent
+                        }
                     }
                     .onAnimationChange(of: translation) { value in
-                        onDrag(value > 0 ? value : 0)
+                        onDrag(value > .zero ? value : .zero)
 
-                        if value <= 0 && sheetConfig?.selectedDetent == .height(.zero) {
-                            isDismissed = true
+                        if (translation == 0 && value == .zero) {
                             isPresented = false
+                            viewWillPresent = false
                         }
                     }
-                    .offset(y: UIScreen.main.bounds.height - translation)
                     .onDisappear {
                         onDismiss()
                     }
@@ -164,22 +163,11 @@ struct SheetPlus<HContent: View, MContent: View, Background: View>: ViewModifier
                         )
                     )
                 }
-                .edgesIgnoringSafeArea([.bottom])
-                .transition(.move(edge: .bottom))
+                .edgesIgnoringSafeArea(.bottom)
+                .zIndex(2)
             }
         }
         .onPreferenceChange(SheetPlusKey.self) { value in
-            /// Quick hack to prevent the scrollview from resetting the height when keyboard shows up.
-            /// Replace if the root cause has been located.
-            if value.detents.count == 0 {
-                isPresented = false
-                return
-            }
-
-            if initialSelectedDetent == nil {
-                initialSelectedDetent = value.selectedDetent
-            }
-
             sheetConfig = value
             translation = value.translation
 
@@ -189,11 +177,43 @@ struct SheetPlus<HContent: View, MContent: View, Background: View>: ViewModifier
         .onPreferenceChange(SheetPlusIndicatorKey.self) { value in
             showDragIndicator = value
         }
-//        .onPreferenceChange(SheetPlusBackgroundInteractionKey.self) { value in
-//            allowBackgroundInteraction = value
-//        }
+        .onPreferenceChange(SheetPlusBackgroundInteractionKey.self) { value in
+            backgroundInteractionDetentLimit = value
+        }
         .onPreferenceChange(SheetPlusInteractiveDismissDisabledKey.self) { value in
             isInteractiveDismissDisabled = value
+        }
+        .onPreferenceChange(SheetPlusPresentationCornerRadiusKey.self) { value in
+            cornerRadius = value
+        }
+        .onPreferenceChange(SheetPlusPresentationBackgroundKey.self) { value in
+            background = value
+        }
+    }
+
+    func onDragEnded(with value: DragGesture.Value) {
+        // Reset the distance on release so we start with a
+        // clean translation next time
+        newValue = 0
+
+        // Calculate velocity based on pt/s so it matches the UIPanGesture
+        let distance: CGFloat = value.translation.height
+        let time: CGFloat = startTime != nil ? value.time.timeIntervalSince(startTime!.time) : 0
+
+        let yVelocity: CGFloat = -1 * ((distance / time) / 1000)
+        startTime = nil
+
+        if let result = snapBottomSheet(
+            translation,
+            detents,
+            yVelocity,
+            isInteractiveDismissDisabled
+        ) {
+            translation = result.size
+
+            if result.size != .zero {
+                sheetConfig?.selectedDetent = result
+            }
         }
     }
 }
