+++ Sources/BottomSheet/Preference Keys/BackgroundInteractionKey.swift
@@ -7,24 +7,32 @@
 
 import SwiftUI
 
-public enum PresentationBackgroundInteractionPlus: Equatable {
-    case automatic
-    case disabled
-    case enabled
+public struct PresentationBackgroundInteractionPlus: Hashable, Sendable {
+    internal enum Kind: Hashable, Sendable {
+        case automatic
+        case disabled
+        case enabled(upThrough: PresentationDetent?)
+    }
 
-    internal static var detent: PresentationDetent?
-    internal static var enabledUpThrough: PresentationBackgroundInteractionPlus = .enabledUpThrough
+    internal let kind: Kind
 
-    public static func enabled(upThrough detent: PresentationDetent) -> PresentationBackgroundInteractionPlus {
-        self.detent = detent
-        return enabledUpThrough
+    internal init(kind: Kind) {
+        self.kind = kind
+    }
+
+    public static let automatic = Self.init(kind: .automatic)
+    public static let enabled = Self.init(kind: .enabled(upThrough: nil))
+    public static let disabled = Self.init(kind: .disabled)
+
+    public static func enabled(upThrough detent: PresentationDetent) -> Self {
+        .init(kind: .enabled(upThrough: detent))
     }
 }
 
 struct SheetPlusBackgroundInteractionKey: PreferenceKey {
-    static var defaultValue: PresentationBackgroundInteractionPlus = .automatic
+    static var defaultValue: CGFloat? = nil
 
-    static func reduce(value: inout PresentationBackgroundInteractionPlus, nextValue: () -> PresentationBackgroundInteractionPlus) {
+    static func reduce(value: inout CGFloat?, nextValue: () -> CGFloat?) {
         value = nextValue()
     }
 }
