+++ Sources/BottomSheet/View Modifiers/View+SheetPlus.swift
@@ -8,14 +8,13 @@
 import SwiftUI
 
 extension View {
-    public func sheetPlus<HContent: View, MContent: View, Background: View>(
+    public func sheetPlus<HContent: View, MContent: View>(
         isPresented: Binding<Bool>,
         animationCurve: SheetAnimation = SheetAnimation(
             mass: SheetAnimationDefaults.mass,
             stiffness: SheetAnimationDefaults.stiffness,
             damping: SheetAnimationDefaults.damping
         ),
-        background: Background = Color(UIColor.systemBackground),
         onDismiss: @escaping () -> Void = {},
         onDrag: @escaping (CGFloat) -> Void = { _ in },
         header: () -> HContent = { EmptyView() },
@@ -23,31 +22,6 @@ extension View {
     ) -> some View {
         modifier(
             SheetPlus(
-                isPresented: isPresented,
-                animationCurve: animationCurve,
-                background: background,
-                onDismiss: onDismiss,
-                onDrag: onDrag,
-                hcontent: header,
-                mcontent: main
-            )
-        )
-    }
-
-    public func sheetPlusV2<HContent: View, MContent: View>(
-        isPresented: Binding<Bool>,
-        animationCurve: SheetAnimation = SheetAnimation(
-            mass: SheetAnimationDefaults.mass,
-            stiffness: SheetAnimationDefaults.stiffness,
-            damping: SheetAnimationDefaults.damping
-        ),
-        onDismiss: @escaping () -> Void = {},
-        onDrag: @escaping (CGFloat) -> Void = { _ in },
-        header: () -> HContent = { EmptyView() },
-        main: () -> MContent
-    ) -> some View {
-        modifier(
-            SheetPlusV2(
                 isPresented: isPresented,
                 animationCurve: animationCurve,
                 onDismiss: onDismiss,
